Overview: Fixed bugs. Details: 1. Now the logics all works. TODO: 1. See if removing goto in Put improve performance. 2. Further test Put and Expand. 3. Add delete.

+++ Maps/HopMap/Element.go
@@ -14,7 +14,7 @@ const (
 type Element[K any, V any] struct {
 	key            K
 	val            V
-	hashOS, linkOS int8
+	hashOS, linkOS int16
 	info           bit
 }
